Add files via upload
This is a basic chat app. a deo really. to open or use it, fork this repo and place the html file and the javascript file in a folder called public and run on your port. http://localhost:1337/static/index.html . port 1337 is the port used locally. navigate to that route and open it on multiple tabs in your browser and voila.